testing/requirements.txt
openai
python-dotenv
fal-client
boto3
opencv-python
numpy
requests
pillow
anthropic
transformers
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
# torch
# torchvision
# torchaudio
# torchaudio